RMG-Core: convert to int when using M64TYPE_BOOL in Settings functions
This is needed for windows, if we don't do this, it seems like there's
some undefined behavior going on, so let's be as explicit as possible.

Weirdly enough the code before this commit works fine on linux, so maybe
some weird platform differences caused this? I'm not sure but being
explicit won't break anything so this should be fine :)