Add Ability to Move and Re-order Columns in "Game List" Mode
In the ROM browser's "Game List" mode, the order of the columns at the top of the screen ("Name", "Internal Name", and "MD5" from left to right by default) can now be changed by users. Doing so is as simple as using the mouse to drag-and-drop one of the column headers. When dropped, the columns will automatically rearrange themselves so that the selected column is moved to where the user dropped it relative to the other two columns.

The layout of the column headers automatically saves itself to the emulator's config file and is retained between sessions. The default layout remains unchanged from previous versions of RMG.