Make Rightmost Column in "Game List" Mode Stretch
When adding the ability to re-order columns in "Game List" mode in my previous commit, I ran into bugs when enabling Qt6's option to stretch the rightmost column, so I disabled it for that commit.

This commit reimplements this feature in a way that doesn't cause any bugs with the new column re-ordering feature. It also makes some behind-the-scenes adjustments to the functions that save and load the column widths to and from the config file, respectively. This was done to ensure that values of "-1" in the config file are always overwritten on startup no matter what, as that was also causing bugs with the new re-ordering feature.

Aside from the rightmost column now always being stretched in "Game List" mode, the user interface should still visibly appear exactly the same as in my previous commit.